Skip Emit assignment when value is null
This commit adds IL instructions to skip the assignment of
null to a type property.

The IL local variable for a class or interface property
will be initialized with null, so if an instance is not deserialized from incoming
JSON, do not call the property setter with null. The property setter may
have side effects when assigning null.