Attributes on interface properties apply to concrete type properties
This commit changes the MetaType and MetaMember functionality
to allow custom attributes on interface properties to be resolved for concrete type properties that
implement the interface. In doing so, this removes the need to specify an JsonFormatterAttribute
on a property of a concrete type where the property is not a library type with its own
JsonFormatter.

Take for example,

IEnumerable<string> CharFilter on ICustomAnalyzer. A CharFilter may be a single value or
an array of values in the response, so is attributed with a special JsonFormatter to handle this.

When deserializing an instance of CustomAnalyzer however, a general collection
Json Formatter will be resolved for this property and not the special JsonFormatter, unless the concrete
type property is also attributed with the special JsonFormatter.